# Add channel metrics support