Adds basic metrics for #1823 

<!-- Please include the 'why' behind your changes if no issue exists -->

## Proposed Changes

- Adds metrics for channels

```
# HELP inmemorychannel_dispatcher_event_count Number of events dispatched by the in-memory channel
# TYPE inmemorychannel_dispatcher_event_count counter
inmemorychannel_dispatcher_event_count{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="202",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d"} 2
# HELP inmemorychannel_dispatcher_event_dispatch_latencies The Time spent dispatching an event from a in-memoryChannel
# TYPE inmemorychannel_dispatcher_event_dispatch_latencies histogram
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="1"} 0
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="2"} 0
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="5"} 0
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="10"} 0
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="20"} 0
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="50"} 0
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="100"} 1
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="200"} 1
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="500"} 1
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="1000"} 1
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="2000"} 1
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="5000"} 2
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="10000"} 2
inmemorychannel_dispatcher_event_dispatch_latencies_bucket{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d",le="+Inf"} 2
inmemorychannel_dispatcher_event_dispatch_latencies_sum{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d"} 2782
inmemorychannel_dispatcher_event_dispatch_latencies_count{container_name="dispatcher",event_type="dev.knative.sources.ping",namespace_name="default",response_code="200",response_code_class="2xx",unique_name="imc-dispatcher-dbd97dcb5-gdr4f75c5f5701783f6e1fcbb3a07ed301dc6d"} 2
```
To test just use:

```
apiVersion: messaging.knative.dev/v1beta1
kind: Channel
metadata:
  name: eventinghello-ch 
---
apiVersion: sources.knative.dev/v1alpha2
kind: PingSource
metadata:
  name: event-greeter-ping-source
spec:
  schedule: "*/1 * * * *"
  jsonData: '{"message": "Hello!!!!"}'
  sink:
   ref:
    apiVersion: messaging.knative.dev/v1beta1
    kind: Channel 
    name: eventinghello-ch
---
apiVersion: serving.knative.dev/v1
kind: Service
metadata:
  name: eventinghelloa
  labels:
    app: secret
spec:
  template:
    metadata:
      name: eventinghelloa-v1 
      annotations:
        autoscaling.knative.dev/target: "1"
    spec:
      containers:
      - image: quay.io/rhdevelopers/eventinghello:0.0.2
---
apiVersion: messaging.knative.dev/v1beta1
kind: Subscription
metadata:
  name: eventinghelloa-sub
spec:
  channel:
    apiVersion: messaging.knative.dev/v1beta1
    kind: Channel
    name: eventinghello-ch
  subscriber:
    ref:
      apiVersion: serving.knative.dev/v1
      kind: Service
      name: eventinghelloa

```
<!--
If this change has user-visible impact, follow the instructions below.
Examples include:

- 🎁 Add new feature
- 🐛 Fix bug
- 🧽 Update or clean up current behavior
- 🗑️ Remove feature or internal logic

Otherwise delete the rest of this template.
-->

**Release Note**

<!--
🗒️ If this change has user-visible impact, write a release note in the block
below. Include the string "action required" if additional action is required of
users switching to the new release, for example in case of a breaking change.

Write as if you are speaking to users, not other Knative contributors. If this
change has no user-visible impact, no release-note is needed.
-->

```release-note

```

**Docs**

<!--
📖 If this change has user-visible impact, link to an issue or PR in
https://github.com/knative/docs.
-->