@slinkydeveloper 2020-10-08
@skonto FYI there's no need to fix the eventing-contrib check