@skonto 2020-10-15
@slinkydeveloper benchstat results look ok:
```
name                                          old time/op    new time/op    delta
pkg:knative.dev/eventing/pkg/inmemorychannel goos:linux goarch:amd64
Dispatcher_dispatch_ok_through_2_channels-12     262µs ± 0%     225µs ± 0%   ~     (p=1.000 n=1+1)
pkg:knative.dev/eventing/test/performance/infra/sender goos:linux goarch:amd64
TargeterFixedPayload-12                          469ns ± 0%     516ns ± 0%   ~     (p=1.000 n=1+1)
TargeterRandomPayload-12                        2.87µs ± 0%    2.60µs ± 0%   ~     (p=1.000 n=1+1)

name                                          old alloc/op   new alloc/op   delta
pkg:knative.dev/eventing/pkg/inmemorychannel goos:linux goarch:amd64
Dispatcher_dispatch_ok_through_2_channels-12    74.7kB ± 0%    77.3kB ± 0%   ~     (p=1.000 n=1+1)
pkg:knative.dev/eventing/test/performance/infra/sender goos:linux goarch:amd64
TargeterFixedPayload-12                           528B ± 0%      528B ± 0%   ~     (all equal)
TargeterRandomPayload-12                          640B ± 0%      640B ± 0%   ~     (all equal)

name                                          old allocs/op  new allocs/op  delta
pkg:knative.dev/eventing/pkg/inmemorychannel goos:linux goarch:amd64
Dispatcher_dispatch_ok_through_2_channels-12       588 ± 0%       666 ± 0%   ~     (p=1.000 n=1+1)
pkg:knative.dev/eventing/test/performance/infra/sender goos:linux goarch:amd64
TargeterFixedPayload-12                           5.00 ± 0%      5.00 ± 0%   ~     (all equal)
TargeterRandomPayload-12                          6.00 ± 0%      6.00 ± 0%   ~     (all equal)
```
I will check again after I update the PR. For the full report check [here](https://gist.github.com/skonto/99bff66002867e7b4c2f4e1c8d529b22).