@skonto 2020-10-15
@slinkydeveloper the benchstat results look ok:
```
name                                          old time/op    new time/op    delta
pkg:knative.dev/eventing/pkg/inmemorychannel goos:linux goarch:amd64
Dispatcher_dispatch_ok_through_2_channels-12     423µs ± 0%     124µs ± 0%   ~     (p=1.000 n=1+1)
pkg:knative.dev/eventing/test/performance/infra/sender goos:linux goarch:amd64
TargeterFixedPayload-12                          973ns ± 0%     775ns ± 0%   ~     (p=1.000 n=1+1)
TargeterRandomPayload-12                        2.49µs ± 0%    2.48µs ± 0%   ~     (p=1.000 n=1+1)

name                                          old alloc/op   new alloc/op   delta
pkg:knative.dev/eventing/pkg/inmemorychannel goos:linux goarch:amd64
Dispatcher_dispatch_ok_through_2_channels-12    75.1kB ± 0%    79.1kB ± 0%   ~     (p=1.000 n=1+1)
pkg:knative.dev/eventing/test/performance/infra/sender goos:linux goarch:amd64
TargeterFixedPayload-12                           528B ± 0%      528B ± 0%   ~     (all equal)
TargeterRandomPayload-12                          640B ± 0%      640B ± 0%   ~     (all equal)

name                                          old allocs/op  new allocs/op  delta
pkg:knative.dev/eventing/pkg/inmemorychannel goos:linux goarch:amd64
Dispatcher_dispatch_ok_through_2_channels-12       589 ± 0%       722 ± 0%   ~     (p=1.000 n=1+1)
pkg:knative.dev/eventing/test/performance/infra/sender goos:linux goarch:amd64
TargeterFixedPayload-12                           5.00 ± 0%      5.00 ± 0%   ~     (all equal)
TargeterRandomPayload-12                          6.00 ± 0%      6.00 ± 0%   ~     (all equal)
```
The full report is [here](https://gist.github.com/skonto/ab0bfa16b51c007be3513ac3332fe03c). I used `./test/microbenchmarks-run.sh` is that enough? I am not sure the [suggestion](https://github.com/knative/eventing/pull/4259#discussion_r504787515) above will have an effect.