@slinkydeveloper 2020-10-15
> pkg:knative.dev/eventing/pkg/inmemorychannel goos:linux goarch:amd64
Dispatcher_dispatch_ok_through_2_channels-12     423µs ± 0%     124µs ± 0%   ~     (p=1.000 n=1+1)

Mhhh this doesn't sound right to me, you should try to run this microbenchmark several times