@skonto 2020-10-15
@slinkydeveloper ok let me try it a few time.