@slinkydeveloper 2020-10-15
@skonto there's an open issue for that https://github.com/knative/test-infra/issues/1906