@slinkydeveloper 2020-10-15
@skonto for eventing-natss you need to talk with @n3wscott, maybe open an issue there

> There is a lot of variance in running time.

I think in the first runs is normal, it's a warmup kind of phase. I think the cpu difference is not a big deal, but the memory usage is consistently higher... I'm gonna try to run locally some profiling on this code to figure out if we can improve it a little bit more