@skonto 2020-10-15
>I think in the first runs is normal, it's a warmup kind of phase. I think the cpu difference is not a big deal, but the memory usage is consistently higher... I'm gonna try to run locally some profiling on this code to figure out if we can improve it a little bit more
I will have a look too, but increase seems to be 5% in terms of bytes. Of course each request will hit this but i am wondering what should be the threshold.