@skonto 2020-10-15
I am hitting: `error: unable to read URL "https://storage.googleapis.com/knative-nightly/serving/latest/monitoring.yaml", server reported 404 Not Found, status code=404
ERROR: Knative Monitoring did not come up`
Is this because the monitoring stuff are being removed from upstream in serving?