@slinkydeveloper 2020-10-19
> Is this because the monitoring stuff are being removed from upstream in serving?

It seems like, can you rebase?