@skonto 2020-10-19
@slinkydeveloper I addressed the comments should I just create the pr for eventing-natss and update the other one for eventing-kafka now?