@vaikas 2020-10-26
@skonto while things are fresh, would you mind tackling this here?
https://github.com/knative/docs/issues/1792